requirements.txt
flask
pandas
python-binance
python-telegram-bot==20.7
python-dotenv
requests